Debug and refactor. Remove unnecessary packages. NOTE: loading a different endpoint after an initial one WILL break the app, so you must refresh before trying a new endpoint. This will be fixed later.
