+++ client/package.json
@@ -11,8 +11,6 @@
     "@graphiql/toolkit": "^0.8.2",
     "@tisoap/react-flow-smart-edge": "^3.0.0",
     "axios": "^0.26.0",
-    "codemirror-graphql": "^2.0.4",
-    "dagre": "^0.8.5",
     "elkjs": "^0.8.2",
     "graphql": "^16.6.0",
     "graphql-request": "^5.2.0",
